+++ HISTORY.rst
@@ -3,12 +3,13 @@ History
 =======
 
 0.2.0 (2021-08-17)
------------------
+------------------
 
 * Replaces setuptools with filt
 * Updates linting
 * Replaces TravisCI with Github Actions
 * Updates for `sxm-player=0.2.1` client
+* Replaces old school Discord commands with proper slash commands
 
 0.1.0 (2018-12-25)
 ------------------